requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
schedule
playwright
apscheduler
